Moved time types to the datetime package
